changed home page, add login, add course save; TODO NEXT:
At the profile page, add a Clear Record, which need admin password
(2025) to click, when click, clear and reset all the course selection
info (for demo purpose)
When login, sync the login info for profile page, when save course, only
save to the loged in student. If I login as another student, the record
should change. If did not login, the profile page will show a warning
for please login